refactor(ui): extract clone manager button into standalone component
